Tabulates into a tibble some properties of the standard atmosphere: temperature, density, and pressure.

Tabulates into a tibble some properties of the standard atmosphere: temperature, density, and pressure.

Usage

atmos_table(units = c("SI", "Eng"), ret_units = TRUE)

Arguments

units

character vector that contains the system of units [options are SI for International System of Units and Eng for English (US customary) units.

ret_units

If set to TRUE the value(s) returned are of class units with units attached to the value. [Default is TRUE]

Author

Ed Maurer

Examples

atmos_table(units = 'SI')
#> # A tibble: 16 × 4
#>    Altitude   Temp Pressure  Density
#>         [m]    [C]     [Pa] [kg/m^3]
#>  1        0  15     101325     1.23 
#>  2     1000   8.50   89876.    1.11 
#>  3     2000   2.00   79501.    1.01 
#>  4     3000  -4.49   70121.    0.909
#>  5     4000 -11.0    61660.    0.819
#>  6     5000 -17.5    54048.    0.736
#>  7     6000 -24.0    47218.    0.660
#>  8     7000 -30.5    41105.    0.590
#>  9     8000 -36.9    35652.    0.526
#> 10     9000 -43.4    30801.    0.467
#> 11    10000 -49.9    26500.    0.414
#> 12    11000 -56.4    22700.    0.365
#> 13    12000 -62.9    19355.    0.321
#> 14    13000 -69.3    16421.    0.281
#> 15    14000 -75.8    13859.    0.245
#> 16    15000 -82.3    11632.    0.212