Solves the Momentum Equation for sequent (or conjugate) depth in a trapezoidal channel

This function solves the Momentum equation for water flow in an open channel with a trapezoidal shape and determines the sequent (conjugate) depth. This is the flow depth either upstream or downstream of a hydraulic jump, whichever is not provided as input.

Usage

sequent_depth(
  Q = NULL,
  b = NULL,
  y = NULL,
  m = NULL,
  units = c("SI", "Eng"),
  ret_units = FALSE
)

Arguments

Q

numeric vector that contains the flow rate [\(m^3 s^{-1}\) or \(ft^3 s^{-1}\)]

b

numeric vector that contains the channel bottom width [\(m\) or \(ft\)]

y

numeric vector that contains the water depth [\(m\) or \(ft\)]

m

numeric vector that contains the side slope of the channel (m:1 H:V) [unitless]

units

character vector that contains the system of units [options are SI for International System of Units and Eng for English (US customary) units. This is used for compatibility with iemisc package.

ret_units

If set to TRUE the value(s) returned are of class units with units attached to the value. [Default is FALSE]

Value

Returns a list including:

y - input depth

y_seq - sequent depth

yc - critical depth

Fr - Froude number for input depth

Fr_seq - Froude number for sequent depth

E - specific energy for input depth

E_seq - specific energy for sequent depth

Details

The Momentum equation for open channel flow conditions in a trapezoidal channel: $$M = \frac{by^2}{2}+\frac{my^3}{3}+\frac{Q^2}{gy\left(b+my\right)}$$ where \(C\) is 1.0 for SI units and 1.49 for Eng (U.S. Customary) units. The momentum function is assumed to be the same on both sides of a hydraulic jump, allowing the determination of the sequent depth.

Author

Ed Maurer

Examples

#Solving for sequent depth: SI Units
#Flow of 0.2 m^3/s, bottom width = 0.5 m, Depth = 0.1 m, side slope = 1:1
sequent_depth(Q=0.2,b=0.5,y=0.1,m=1,units = "SI", ret_units = TRUE)
#> Mixed units: 1 (2), m (5) 
#> 0.1 [m], 0.3941009 [m], 0.217704 [m], 3.635731 [1], 0.3465538 [1], 0.666509 [m], 0.4105265 [m]