Tabulates into a tibble the basic water properties: density, dynamic and kinematic viscosity, saturation vapor pressure, surface tension, and bulk modulus.

Tabulates into a tibble the basic water properties: density, dynamic and kinematic viscosity, saturation vapor pressure, surface tension, and bulk modulus.

Usage

water_table(units = c("SI", "Eng"), ret_units = TRUE)

Arguments

units

character vector that contains the system of units [options are SI for International System of Units and Eng for English (US customary) units. This is used for compatibility with iemisc package

ret_units

If set to TRUE the value(s) returned are of class units with units attached to the value. [Default is TRUE]

Author

Ed Maurer

Examples

water_table(units = 'SI')
#> # A tibble: 21 × 8
#>    Temperature  Density Spec_Weight Viscosity  Kinem_Visc Sat_VP Surf_…¹ Bulk_…²
#>            [C] [kg/m^3]     [N/m^3] [N*s/m^2]     [m^2/s]   [Pa]   [N/m]    [Pa]
#>  1           0    1000.       9809.  0.00173  0.00000173    611.  0.0757  2.02e9
#>  2           5    1000.       9810.  0.00150  0.00000150    873.  0.0749  2.06e9
#>  3          10    1000.       9807.  0.00131  0.00000131   1228.  0.0742  2.1 e9
#>  4          15     999.       9801.  0.00115  0.00000115   1706.  0.0735  2.14e9
#>  5          20     998.       9793.  0.00102  0.00000102   2339.  0.0727  2.18e9
#>  6          25     997.       9781.  0.000911 0.000000913  3170.  0.0720  2.22e9
#>  7          30     996.       9768.  0.000817 0.000000821  4247.  0.0712  2.25e9
#>  8          35     994.       9752.  0.000738 0.000000742  5629.  0.0704  2.26e9
#>  9          40     992.       9734.  0.000670 0.000000675  7385.  0.0696  2.28e9
#> 10          45     990.       9714.  0.000611 0.000000617  9595.  0.0688  2.28e9
#> # … with 11 more rows, and abbreviated variable names ¹​Surf_Tens, ²​Bulk_Mod