Functions to calculate water properties: density, dynamic and kinematic viscosity, saturation vapor pressure, surface tension, and bulk modulus.

This function calculates water properties that are used in other functions.

Usage

dvisc(T = NULL, units = c("SI", "Eng"), ret_units = FALSE)

dens(T = NULL, units = c("SI", "Eng"), ret_units = FALSE)

kvisc(T = NULL, units = c("SI", "Eng"), ret_units = FALSE)

svp(T = NULL, units = c("SI", "Eng"), ret_units = FALSE)

surf_tension(T = NULL, units = c("SI", "Eng"), ret_units = FALSE)

Ev(T = NULL, units = c("SI", "Eng"), ret_units = FALSE)

Arguments

T

the water temperature [\(^{\circ}C\) or \(^{\circ}F\)]

units

character vector that contains the system of units [options are SI for International System of Units and Eng for English (US customary) units. This is used for compatibility with iemisc package

ret_units

If set to TRUE the value(s) returned are of class units with units attached to the value. [Default is FALSE]

Value

rho, the density of water for the dens function [\({kg}\,{m^{-3}}\) or \({slug}\,{ft^{-3}}\)] mu, the dynamic viscosity of water for the dvisc function [\({N}\,{s}\,{m^{-2}}\) or \({lbf}\,{s}\,{ft^{-2}}\)] nu, the kinematic viscosity of water for the kvisc function [\(m^2 s^{-1}\) or \(ft^2 s^{-1}\)]. svp, the saturation vapor pressure of water for the svp function [\(N m^{-2}\) or \(lbf ft^{-2}\)]. surf_tension, the surface tension of water for the surf_tension function [\(N m^{-1}\) or \(lbf ft^{-1}\)]. Ev, the bulk modulus of elasticity of water for the Ev function [\(N m^{-2}\) or \(lbf ft^{-2}\)].

Author

Ed Maurer

Examples

#Find kinematic viscocity for water temperature of 55 F
nu = kvisc(T = 55, units = 'Eng')

#Find kinematic viscocity assuming default water temperature of 68 F
nu = kvisc(units = 'Eng')
#> 
#> Temperature not given.
#> Assuming T = 68 F

#Find water density for water temperature of 25 C
rho = dens(T = 25, units = 'SI')

#Find saturation vapor pressure for water temperature of 10 C
vps = svp(T = 10, units = 'SI')

#Find surface tension for water temperature of 10 C
s_tens = surf_tension(T = 10, units = 'SI')