Tabulates into a tibble the basic water properties: density, dynamic and kinematic viscosity, saturation vapor pressure, surface tension, and bulk modulus.

Tabulates into a tibble the basic water properties: density, dynamic and kinematic viscosity, saturation vapor pressure, surface tension, and bulk modulus.

Usage

water_table(units = c("SI", "Eng"), ret_units = TRUE)

Arguments

units

character vector that contains the system of units [options are SI for International System of Units and Eng for English (US customary) units. This is used for compatibility with iemisc package

ret_units

If set to TRUE the value(s) returned are of class units with units attached to the value. [Default is TRUE]

Author

Ed Maurer

Examples

water_table(units = 'SI')
#> # A tibble: 21 × 8
#>    Temp  Density Spec_Weight Viscosity  Kinem_Visc Sat_VP Surf_Tens   Bulk_Mod
#>     [C] [kg/m^3]     [N/m^3] [N*s/m^2]     [m^2/s]   [Pa]     [N/m]       [Pa]
#>  1    0    1000.       9809.  0.00173  0.00000173    611.    0.0757 2020000000
#>  2    5    1000.       9810.  0.00150  0.00000150    873.    0.0749 2060000000
#>  3   10    1000.       9807.  0.00131  0.00000131   1228.    0.0742 2100000000
#>  4   15     999.       9801.  0.00115  0.00000115   1706.    0.0735 2140000000
#>  5   20     998.       9793.  0.00102  0.00000102   2339.    0.0727 2180000000
#>  6   25     997.       9781.  0.000911 0.000000913  3170.    0.0720 2220000000
#>  7   30     996.       9768.  0.000817 0.000000821  4247.    0.0712 2250000000
#>  8   35     994.       9752.  0.000738 0.000000742  5629.    0.0704 2265000000
#>  9   40     992.       9734.  0.000670 0.000000675  7385.    0.0696 2280000000
#> 10   45     990.       9714.  0.000611 0.000000617  9595.    0.0688 2285000000
#> # ℹ 11 more rows